requirements.txt
docx2txt==0.8
fastapi==0.115.8
Flask==3.1.0
flask_cors==5.0.0
flask_jwt_extended==4.7.1
flask_sqlalchemy==3.1.1
psycopg2==2.9.1
magika==0.5.1
numpy>=1.26,<2.0
openai==1.61.1
opencv_python==4.11.0.86
passlib==1.7.4
Pillow==11.1.0
pydantic==2.10.6
PyJWT==2.8.0
PyPDF2==3.0.1
python-dotenv==1.0.1
rapid_table_det==1.0.3
Requests==2.32.3
SQLAlchemy==2.0.25
uvicorn==0.34.0
Werkzeug==3.1.3
waitress
passlib